File group management repo service controller

+++ src/main/java/com/example/enterprise_internet_applications_project/services/FilesService.java
@@ -23,7 +23,9 @@ public class FilesService {
         filesRepository.save(fileDB);
     }
 
-    public MyFile getFile(int id) {
+    public MyFile getFile(Long id) {
         return filesRepository.findById(id).get();
     }
+
+
 }
